backend/requirements.txt
flask==2.3.2
psycopg2-binary==2.9.6
werkzeug==2.3.6
pyjwt==2.7.0
flask-cors==3.0.10